refactor: Migrate onboarding template in a Gradle module.
