+++ app/build.gradle.kts
@@ -78,6 +78,7 @@ dependencies {
     implementation(project(":demos:paint"))
     implementation(project(":demos:cryptoapp:app"))
     implementation(project(":demos:moviesapp:app"))
+    implementation(project(":templates:onboarding"))
     implementation(project(":components:fab"))
     implementation(project(":components:charts"))
     implementation(project(":components:tags"))
