+++ settings.gradle.kts
@@ -23,3 +23,4 @@ include(":demos:datingapp")
 include(":animations:lottie")
 include(":animations:canvas")
 include(":components:comingsoon")
+include(":templates:onboarding")
