+++ templates/onboarding/.gitignore
@@ -1,0 +1,1 @@
+/build
