+++ templates/onboarding/build.gradle.kts
@@ -1,0 +1,15 @@
+import com.guru.composecookbook.build.dependencies.*
+
+plugins {
+    /**
+     * See [common-compose-module-configs-script-plugin.gradle.kts] file
+     */
+    id("common-compose-module-configs-script-plugin")
+}
+
+dependencies {
+    implementation(project(":components:carousel"))
+    implementation(project(":animations:lottie"))
+
+    addComposeOfficialDependencies()
+}
