+++ templates/onboarding/src/main/java/com/guru/composecookbook/onboarding/OnboardingPagerItem.kt
@@ -1,0 +1,44 @@
+package com.guru.composecookbook.onboarding
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.guru.composecookbook.lottie.LottieLoadingView
+
+@Composable
+fun OnboardingPagerItem(item: Onboard) {
+    Column(
+        modifier = Modifier.padding(horizontal = 16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        LottieLoadingView(
+            context = LocalContext.current,
+            file = item.lottieFile,
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(250.dp)
+        )
+        Text(
+            text = item.title,
+            style = MaterialTheme.typography.h5.copy(fontWeight = FontWeight.ExtraBold),
+            textAlign = TextAlign.Center,
+            modifier = Modifier.padding(16.dp)
+        )
+        Text(
+            text = item.description,
+            style = MaterialTheme.typography.body1,
+            textAlign = TextAlign.Center,
+            modifier = Modifier.padding(horizontal = 16.dp)
+        )
+    }
+}
