+++ templates/onboarding/src/main/java/com/guru/composecookbook/onboarding/OnboardingPagerSlide.kt
@@ -1,0 +1,19 @@
+package com.guru.composecookbook.onboarding
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun OnboardingPagerSlide(selected: Boolean, color: Color, icon: ImageVector) {
+    Icon(
+        imageVector = icon,
+        modifier = Modifier.padding(4.dp),
+        contentDescription = null,
+        tint = if (selected) color else Color.Gray
+    )
+}
